update(infra): types different file for use in functions directly
need to separate it out so that imports in function don't import all the
other sst specific stuff
