+++ infra/builder/step.ts
@@ -1,15 +1,4 @@
-export const ARM_INSTANCE_TYPES = ['c7g.16xlarge', 'c8g.16xlarge', 'c8g.48xlarge'] as const;
-export const X86_INSTANCE_TYPES = [
-  'c7i.16xlarge',
-  'c7i.24xlarge',
-  'c8i.24xlarge',
-  'c7i.metal-48xl',
-  'm8i.24xlarge',
-  'c7i.48xlarge',
-  'c8a.48xlarge',
-  'c8i.48xlarge',
-  'c8i.metal-48xl'
-] as const;
+import { ARM_INSTANCE_TYPES, X86_INSTANCE_TYPES } from './types';
 
 function launchInstance(architecture: 'x86' | 'arm64', market: 'spot' | 'on-demand') {
   return {
@@ -38,7 +27,6 @@ function launchInstance(architecture: 'x86' | 'arm64', market: 'spot' | 'on-dema
           ResourceType: 'instance',
           Tags: [
             { Key: 'BuildId', 'Value.$': '$.buildId' },
-            { Key: 'Script', 'Value.$': '$.script' },
             { Key: 'RepoRef', 'Value.$': '$.repoRef' }
           ]
         }
